# split read done channel from bufchan