--- a/pkg/stream/http/stream_test.go
+++ b/pkg/stream/http/stream_test.go
@@ -18,6 +18,7 @@
 package http
 
 import (
+	"bufio"
 	"bytes"
 	"context"
 	"fmt"
@@ -77,6 +78,42 @@ func Test_clientStream_AppendHeaders(t *testing.T) {
 	}
 }
 
+func TestStreamConnectionDispatch(t *testing.T) {
+	streamConnectionMocked := &streamConnection{
+		bufChan:    make(chan buffer.IoBuffer),
+		endRead:    make(chan struct{}),
+		connClosed: make(chan bool, 1),
+	}
+	streamConnectionMocked.br = bufio.NewReaderSize(streamConnectionMocked, defaultMaxHeaderSize)
+	httpTestResponseHeader := "HTTP/1.1 200 OK\r\nDate: Fri, 13 Nov 2020 09:27:39 GMT\r\nContent-Type: text/plain; charset=utf-8\r\nContent-Length: 12\r\n\r\n"
+	httpTestResponseBody := `hello`
+	httpTestResponseBody2 := ` world!`
+	go streamConnectionMocked.Dispatch(buffer.NewIoBufferString(httpTestResponseHeader))
+	// wait Dispatch ready
+	time.Sleep(time.Second)
+	go streamConnectionMocked.Dispatch(buffer.NewIoBufferString(httpTestResponseBody))
+
+	time.Sleep(time.Second)
+	go streamConnectionMocked.Dispatch(buffer.NewIoBufferString(httpTestResponseBody2))
+
+	response := fasthttp.AcquireResponse()
+	// wait Dispatch ready
+	time.Sleep(time.Second)
+	err := response.Read(streamConnectionMocked.br)
+	if err != nil {
+		t.Fatalf("http reponse read error: %v", err)
+	}
+
+	//t.Logf("Header: %v body: %v", response.Header.String(), string(response.Body()))
+	if response.Header.ContentLength() != 12 {
+		t.Errorf("want length: %v get: %v", 12, response.Header.ContentLength())
+	}
+
+	if string(response.Body()) != httpTestResponseBody+httpTestResponseBody2 {
+		t.Errorf("want body: %v. get: %v", httpTestResponseBody+httpTestResponseBody2, string(response.Body()))
+	}
+}
+
 func Test_header_capitalization(t *testing.T) {
 	remoteAddr, _ := net.ResolveTCPAddr("tcp", "127.0.0.1:12200")
 
@@ -102,7 +139,7 @@ func Test_header_capitalization(t *testing.T) {
 		{
 			protocol.MosnHeaderQueryStringKey: queryString,
 			protocol.MosnHeaderPathKey:        path,
-			"Args": "Hello, world!",
+			"Args":                            "Hello, world!",
 		},
 	}
 
@@ -237,7 +274,7 @@ func Test_serverStream_handleRequest(t *testing.T) {
 		name   string
 		fields fields
 	}{
-	// TODO: Add test cases.
+		// TODO: Add test cases.
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {