@alpha-baby 2020-11-13
buf 复现：

```log
--- FAIL: TestStreamConnectionDispatch (3.32s)
    stream_test.go:113: want body: hello world!. get:  world!hello
FAIL
```

我些的这个单元测试如果想要比较好的复现需要修改下 read函数：

```go
func (conn *streamConnection) Read(p []byte) (n int, err error) {
	// conn.bufChan <- nil Maybe caused panic error when connection closed.
	defer func() {
		if r := recover(); r != nil {
			log.DefaultLogger.Infof("[stream] [http] connection has closed. Connection = %d, Local Address = %+v, Remote Address = %+v, err = %+v",
				conn.conn.ID(), conn.conn.LocalAddr(), conn.conn.RemoteAddr(), r)
		}
	}()

	data, ok := <-conn.bufChan
	time.Sleep(time.Millisecond * 100) //tip 在这里添加上超时
	// Connection close
	if !ok {
		// Compatible with the fasthttp,
		// it should be set err = IO.EOF when the peer connection is closed.
		switch conn.resetReason {
		case types.UpstreamReset:
			err = io.EOF
		default:
			err = errConnClose
		}

		return
	}

	n = copy(p, data.Bytes())
	data.Drain(n)
	conn.bufChan <- nil
	return
}
```

`Dispatch`函数确实不会丢失数据，但是在`Dispatch`函数中会出现长时间的自旋，然后导致数据不一致

如果想要看到自旋的问题再到此函数中加入日志打印就可以：

```go
func (sc *streamConnection) Dispatch(buffer buffer.IoBuffer) {
	defer func() {
		if r := recover(); r != nil {
			log.DefaultLogger.Errorf("[stream] [http] connection has closed. Connection = %d, Local Address = %+v, Remote Address = %+v, err = %+v",
				sc.conn.ID(), sc.conn.LocalAddr(), sc.conn.RemoteAddr(), r)
		}
	}()

	for buffer.Len() > 0 {
		sc.bufChan <- buffer
		<-sc.bufChan
		log.DefaultLogger.Infof("len: %v", buffer.Len())
	}
	log.DefaultLogger.Infof("exit: %v", buffer.String())
}
```

最新 commit 修此问题，unit test 结果如下：

```log
=== RUN   TestStreamConnectionDispatch
2020-11-13 18:30:35,196 [INFO] len: 0
2020-11-13 18:30:35,196 [INFO] exit: 
2020-11-13 18:30:35,297 [INFO] len: 0
2020-11-13 18:30:35,297 [INFO] exit: 
--- PASS: TestStreamConnectionDispatch (3.31s)
2020-11-13 18:30:35,398 [INFO] len: 0
2020-11-13 18:30:35,398 [INFO] exit: 
PASS
```