@alpha-baby 2020-11-13
例如在我本地出现的自旋，输出的日志如下：

```log
2020-11-13 18:12:43,970 [INFO] len: 117
2020-11-13 18:12:43,970 [INFO] len: 5
2020-11-13 18:12:44,421 [INFO] len: 5
2020-11-13 18:12:44,803 [INFO] len: 117
2020-11-13 18:12:44,803 [INFO] len: 7
2020-11-13 18:12:45,196 [INFO] len: 5
2020-11-13 18:12:45,738 [INFO] len: 0
2020-11-13 18:12:45,738 [INFO] exit: 
2020-11-13 18:12:46,119 [INFO] len: 5
2020-11-13 18:12:46,467 [INFO] len: 7
2020-11-13 18:12:46,785 [INFO] len: 5
2020-11-13 18:12:47,131 [INFO] len: 7
2020-11-13 18:12:47,131 [INFO] len: 0
2020-11-13 18:12:47,131 [INFO] exit: 
2020-11-13 18:12:48,420 [INFO] len: 0
2020-11-13 18:12:48,420 [INFO] exit: 
// ...
// ....
```