@taoyuanyuan 2020-11-16
@alpha-baby 你的test是不对的，一个连接同时最多只会一个协程调用Dispatch，不会并发的。