@alpha-baby 2020-11-16
是因为在http1.X 协议汇总，一个TCP连接在同一个时间只会发送一个请求，是这样理解吗？