@alpha-baby 2020-11-16
我又看看下代码，确实是这样的，
如果是这样的话，我觉得我这样修改也可以增强代码的可读性和健壮性，
如果你们觉得没得必要的话，就把这个PR关掉就行了
@taoyuanyuan