@cch123 2020-12-02
感觉可以，需要试一下 http 的性能有没有影响，用 example 里的 demo 和 wrk 简单测一下 post 上来就行，
如果差不多的话下版本合并~