@alpha-baby 2020-12-05
感觉可以，需要试一下 http 的性能有没有影响，用 example 里的 demo 和 wrk 简单测一下 post 上来就行，
如果差不多的话下版本合并~

test environment： MacBook Pro (16-inch, 2019) 2.6 GHz 六核Intel Core i7 16 GB 2667 MHz DDR4

start http-sample: GOMAXPROCS=4 go run ./examples/codes/http-sample/server.go