@alpha-baby 2020-12-05
> 感觉可以，需要试一下 http 的性能有没有影响，用 example 里的 demo 和 wrk 简单测一下 post 上来就行，
> 如果差不多的话下版本合并~

test environment： MacBook Pro (16-inch, 2019) 2.6 GHz 六核Intel Core i7 16 GB 2667 MHz DDR4

start http-sample: `GOMAXPROCS=4 go run ./examples/codes/http-sample/server.go`

### stress result without proxy:

```
$ wrk -t4 -c500 -d30s --latency http://127.0.0.1:8080/test-for-example-http
Running 30s test @ http://127.0.0.1:8080/test-for-example-http
  4 threads and 500 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     2.48ms  744.02us  22.21ms   82.53%
    Req/Sec    25.16k     4.25k   33.36k    75.75%
  Latency Distribution
     50%    2.33ms
     75%    2.78ms
     90%    3.29ms
     99%    5.07ms
  3003886 requests in 30.01s, 1.41GB read
  Socket errors: connect 251, read 91, write 0, timeout 0
Requests/sec: 100110.17
Transfer/sec:     48.02MB
``` 

### strat mosn
```bash
$ make build-local && GOMAXPROCS=4 ./build/bundles/v0.18.0/binary/mosnd -c ./examples/codes/http-sample/server_config.json
```

### stress result with proxy using mosn 

**before optimization**

```
$ wrk -t4 -c500 -d30s --latency http://127.0.0.1:2046/test-for-example-http
Running 30s test @ http://127.0.0.1:2046/test-for-example-http
  4 threads and 500 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     8.20ms    1.80ms  31.72ms   81.12%
    Req/Sec     7.62k   738.27     9.66k    71.17%
  Latency Distribution
     50%    7.89ms
     75%    8.77ms
     90%   10.20ms
     99%   14.79ms
  909566 requests in 30.01s, 452.80MB read
  Socket errors: connect 251, read 98, write 0, timeout 0
Requests/sec:  30304.52
Transfer/sec:     15.09MB
```

**after optimization**

```
$ wrk -t4 -c500 -d30s --latency http://127.0.0.1:2046/test-for-example-http
Running 30s test @ http://127.0.0.1:2046/test-for-example-http
  4 threads and 500 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     7.91ms    1.80ms  46.97ms   79.42%
    Req/Sec     7.90k     0.87k   10.77k    72.58%
  Latency Distribution
     50%    7.67ms
     75%    8.56ms
     90%    9.90ms
     99%   13.85ms
  943180 requests in 30.01s, 469.53MB read
  Socket errors: connect 251, read 99, write 0, timeout 0
Requests/sec:  31431.60
Transfer/sec:     15.65M
```